draw.py refactor
There were several objectives for these changes:
1. Allow dividers to be placed anywhere on a page, in any ordinal
rotation.  This includes flipping individual dividers from the front of
the page to the back of the page.
2. Improve the layout efficiency of the dividers on the page.  The added
improvements include: (a) checking to see if rotating the dividers
allows more to be printed on a page, and (b) checking to see if there is
enough room to add one extra set of cards in another rotation.